prototype/requirements.txt
grpcio==1.64.0
pyroute2==0.7.12
scapy==2.5.0
